search: product search with filtering

+++ src/main/java/pl/pongut/warehouse/data/product/Product.java
@@ -10,7 +10,6 @@ import org.springframework.data.annotation.TypeAlias;
 import org.springframework.data.mongodb.core.mapping.Document;
 import pl.pongut.warehouse.data.supplier.Supplier;
 
-import java.math.BigDecimal;
 
 @Document(collection = "products")
 @TypeAlias("product")
@@ -24,7 +23,7 @@ public class Product {
     private ObjectId _id;
     private String productName;
     private Integer quantityPerUnit;
-    private BigDecimal unitPrice;
+    private Double unitPrice;
     private Integer unitsInStock;
     private Boolean discount;
     private String categoryName;
